Added LCD; housekeeping
Added supporting code for the LCD. It should be standalone now, with serial as an optional 'extra data' option.
Removed the serial slow mode toggle. The pins are needed for the LCD, and it somewhat obsoletes the slow feature anyway.
Generally cleaned up the code for testing. Hopefully the final revision!